Direction added
