+++ headers/types.hpp
@@ -33,41 +33,11 @@ namespace chess {
         return Color(static_cast<uint8_t>(c) ^ 1);
     }
 
-    enum class Direction : int8_t { 
-        NORTH = 8, 
-        WEST = -1, 
-        SOUTH = -8, 
-        EAST = 1,
-        NORTH_EAST = 9, 
-        NORTH_WEST = 7, 
-        SOUTH_WEST = -9, 
-        SOUTH_EAST = -7 
-    };
-
-    template <Color c>
-    constexpr Direction relativeDirection(Direction d) {
-        if constexpr (c == Color::WHITE) {
-            return d;
-        } else {
-            return static_cast<Direction>(-static_cast<int8_t>(d));
-        }
-    }
-
-    template <Color c, Direction d>
-    constexpr Direction relativeDirection() {
-        if constexpr (c == Color::WHITE) {
-            return d;
-        } else {
-            return static_cast<Direction>(-static_cast<int8_t>(d));
-        }
-    }
-
     constexpr uint8_t NUM_SQUARES       = 64;
     constexpr uint8_t NUM_COLORS        = 2;
     constexpr uint8_t NUM_PIECES        = 12;
     constexpr uint8_t NUM_PIECE_TYPES   = 6;
     constexpr int     MAX_MOVES         = 256;
-    constexpr U64     DEFAULT_CHECKMASK = 0xFFFFFFFFFFFFFFFF;
 
     constexpr Piece charToPiece(char c){
